refactor(tests): split common/sweep.rs into bm25/hybrid submodules (#98)
PR #97 consolidated four near-duplicate retrieval-sweep harnesses into a
single 557-line tests/common/sweep.rs, over the project's 500-line
guideline. Split it into a directory module:

- sweep/mod.rs   — declares submodules; hybrid is fastembed-gated
- sweep/bm25.rs  — BM25 grid, selection, and characterization tests
- sweep/hybrid.rs — hybrid grid, selection, and tests

Shims now import from the submodule path directly
(common::sweep::{bm25,hybrid}::{...}), which lets the
`#[allow(unused_imports)] pub use hybrid::{...}` re-export go away — its
sole reason for existing was that every test binary pulling in
`mod common;` compiled the re-export but most never used it.

Inline cleanups on the moved code: dropped a now-unused KnowledgeBase
import in the hybrid module, and silenced a pre-existing
clippy::too_many_arguments on evaluate_hybrid_cell (never gated in CI).

Behavior and printed-table output are unchanged. Each file is now under
the 500-line guideline (370 / 378 / 34).

Verified: cargo test -p primer-kb-load (14 passed); hybrid selection
tests under --features fastembed (10 passed); clippy clean on the sweep
files; fmt clean.
